+++ docs/CHANGELOG.rst
@@ -6,7 +6,7 @@ All notable changes to the Vector SIL Kit project shall be documented in this fi
 The format is based on `Keep a Changelog (http://keepachangelog.com/en/1.0.0/) <http://keepachangelog.com/en/1.0.0/>`_.
 
 
-[4.0.23] - UNRELEASED
+[4.0.23] - 2023-04-17
 ---------------------
 
 Changed
@@ -21,6 +21,12 @@ Fixed
 
 - Ethernet and CAN (Trivial Sim.): The self-delivery with ``TransmitDirection::TX`` is now triggered only after
   the frame has been sent with ``TransmitDirection::RX``.
+
+
+Removed
+~~~~~~~
+
+- Visual Studio 2015 is no longer maintained and therefore not officially supported anymore.
 
 
 [4.0.22] - 2023-04-05